+++ src/auth/ValClient/CookieAuth.ts
@@ -7,13 +7,16 @@ import { AuthFlow } from "./AuthFlow";
 
 import type { IValClient_Auth } from "../../resources/interface/IValClient";
 import type { IAxiosClient } from '../../resources/interface/IAxiosClient';
+import { Axios, AxiosResponse } from 'axios';
 
 //class
 
 /**
- * * Class ID: @ing3kth/val-api/ValClient/Multifactor
+ * *Not Recommend*
+ * 
+ * * Class ID: @ing3kth/val-api/ValClient/CookieAuth
  */
-class Multifactor {
+class CookieAuth {
     public classId:string;
     private cookie:CookieJar;
     private access_token:string;
@@ -31,11 +34,11 @@ class Multifactor {
     * @param {IValClient_Auth} data Account toJSON data
     */
     constructor(data: IValClient_Auth) {
-        if(!data.multifactor){
-            IngCore.Logs.log('This Account is not have a Multifactor', 'error', true);
+        if(data.multifactor){
+            IngCore.Logs.log('This Account is have a Multifactor', 'error', true);
         }
 
-        this.classId = '@ing3kth/val-api/ValClient/Multifactor';
+        this.classId = '@ing3kth/val-api/ValClient/CookieAuth';
         this.cookie = CookieJar.fromJSON(JSON.stringify(data.cookie));
         this.access_token = data.access_token;
         this.id_token = data.id_token;
@@ -47,31 +50,28 @@ class Multifactor {
     }
 
     /**
-    * @param {Number} verificationCode Verification Code
-    * @returns {Promise<IValClient_Auth>}
+    * @returns {Promise<any>}
     */
-    public async execute(verificationCode:number):Promise<IValClient_Auth> {
-        const axiosClient:AxiosClient = new AxiosClient({
+    public async execute():Promise<any> {
+        const axiosClient:Axios = new AxiosClient({
             jar: this.cookie,
             withCredentials: true,
             headers: {
                 'User-Agent': IngCore.Config["val-api"].ValClient.auth["User-Agent"],
             },
-            timeout: this.expires_in * 1000,
-        });
+            maxRedirects: 1,
+        }).axiosClient;
 
-        //ACCESS TOKEN
-        const auth_response:IAxiosClient = await axiosClient.put('https://auth.riotgames.com/api/v1/authorization', {
-            "type": "multifactor",
-            "code": verificationCode.toString(),
-            "rememberDevice": true
-        });
-
-        if(!auth_response.isError){
-            this.multifactor = false;
+        //Cookie Reauth
+        var reauth_response:AxiosResponse;
+        try{
+            reauth_response = await axiosClient.get('https://auth.riotgames.com/authorize?redirect_uri=https%3A%2F%2Fplayvalorant.com%2Fopt_in&client_id=play-valorant-web-prod&response_type=token%20id_token&nonce=1')
+        }catch(err:any){
+            return await AuthFlow.fromUrl(this.toJSON(), err.request._currentUrl as string);
         }
 
-        return await AuthFlow.execute(this.toJSON(), auth_response);
+        IngCore.Logs.log(this.classId + ' URL not find', 'error', true);
+        return this.toJSON();
     }
 
     /**
@@ -95,14 +95,13 @@ class Multifactor {
 
     /**
     * @param {IValClient_Auth} data ValAuth_Account toJSON data
-    * @param {Number} verificationCode Verification Code
     * @returns {Promise<IValClient_Auth>}
     */
-    static async verify(data:IValClient_Auth, verificationCode:number):Promise<IValClient_Auth> {
-        const MultifactorAccount:Multifactor = new Multifactor(data);
-        return await MultifactorAccount.execute(verificationCode);
+     static async reauth(data:IValClient_Auth):Promise<IValClient_Auth> {
+        const CookieAccount:CookieAuth = new CookieAuth(data);
+        return await CookieAccount.execute();
     }
 }
 
 //export
-export { Multifactor };
+export { CookieAuth };