+++ src/service/RiotApi/ContentV1.ts
@@ -13,10 +13,10 @@ import _Locale from "../../resources/data/Locale";
  * * Class ID: @ing3kth/val-api/RiotApi/ContentV1
  */
 class ContentV1 {
-    classId:string;
-    apiKey:string;
-    region:IValRegion;
-    AxiosClient:AxiosClient;
+    public classId:string;
+    private apiKey:string;
+    private region:IValRegion;
+    private AxiosClient:AxiosClient;
 
     /**
     * @param {JSON} data Services Data
@@ -34,7 +34,7 @@ class ContentV1 {
      * @param {String} locale Locale
      * @returns {Promise<IAxiosClient>}
      */
-    async Contents(locale:keyof typeof _Locale.data = 'English_United_States'):Promise<IAxiosClient> {
+     public async Contents(locale:keyof typeof _Locale.data = 'English_United_States'):Promise<IAxiosClient> {
         return await this.AxiosClient.get(this.region.riot.server + `/val/content/v1/contents?locale=${_Locale.data[locale]}` + `&api_key=${this.apiKey}`);
     }
 }